moved ISettingsService to core and implementation to ConsoleBotHost
