+++ Extensions/GithubAnnouncements/packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Analects.SettingsService" version="1.0" />
   <package id="Newtonsoft.Json" version="4.0.5" />
 </packages>
