+++ Jabbot.ConsoleBotHost/Jabbot.ConsoleBotHost.csproj
@@ -86,6 +86,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="SettingsService.cs" />
     <Compile Include="IScheduler.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
