+++ Jabbot.ConsoleBotHost/Scheduler.cs
