+++ Jabbot.ConsoleBotHost/app.config
@@ -14,12 +14,9 @@
   </appSettings>
   <system.diagnostics>
     <sources>
-      <source name="System.ComponentModel.Composition"
-      switchValue="All">
+      <source name="System.ComponentModel.Composition" switchValue="All">
         <listeners>
-          <add name="fileListener"
-          type="System.Diagnostics.TextWriterTraceListener"
-          initializeData="composition.log" />
+          <add name="fileListener" type="System.Diagnostics.TextWriterTraceListener" initializeData="composition.log" />
         </listeners>
       </source>
     </sources>
