+++ Jabbot/Jabbot.csproj
@@ -54,6 +54,7 @@
   <ItemGroup>
     <Compile Include="Bot.cs" />
     <Compile Include="BotExtensions.cs" />
+    <Compile Include="ISettingsService.cs" />
     <Compile Include="Models\ChatMessage.cs" />
     <Compile Include="Models\ChatUser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
